# speed up check for single replicate in `calculate()`