--- a/R/calculate.R
+++ b/R/calculate.R
@@ -274,7 +274,8 @@ calc_impl_one_f <- function(f) {
       dplyr::summarize(stat = f(!!(sym(col)), ...))
 
     # calculate SE for confidence intervals
-    if (length(unique(x[["replicate"]])) == 1) {
+    if ("replicate" %in% dplyr::group_vars(x) &&
+        nrow(dplyr::group_keys(x)) == 1) {
       sample_sd <- x %>%
         dplyr::summarize(stats::sd(!!(sym(col)))) %>%
         dplyr::pull()