This `unique()` calculation is especially burdensome with larger datasets and many replicates--that information is stored in the attributes of `x` already. A ~25% speedup here, affecting `stat %in% c("mean", "median", "sum", "sd")`.

With `main` dev:

``` r
library(infer)

set.seed(1)

gss_gen <-
   gss %>%
   specify(response = hours) %>%
   hypothesize(null = "point", mu = 40) %>%
   generate(reps = 1000, type = "bootstrap")

bench::mark(
    calculate = calculate(gss_gen, stat = "mean")
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 calculate    12.5ms   13.1ms      76.8    14.9MB     68.2
```

With this PR:

``` r
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 calculate    9.31ms   9.55ms      104.    9.01MB     50.2
```

<sup>Created on 2023-04-10 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>